requirements.txt
annotated-types==0.7.0
asarPy==1.0.1
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
defusedxml==0.7.1
fonttools==4.61.1
fpdf2==2.8.6
greenlet==3.3.1
idna==3.11
markdownify==0.13.1
numpy==1.26.3
pandas==2.3.3
pillow==12.1.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jobspy==1.1.82
pytz==2025.2
regex==2024.11.6
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
tls-client==1.0.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
